Move machine parsing logic to a separate file
